requirements.txt
Flask==3.1.1
python-dotenv==1.1.0
openai==1.79.0
requests==2.32.3
Flask-RESTX>=1.3.0
holidays
tweepy>=4.10.0
